Refactors FollowUpDto
