+++ sormas-api/src/main/java/de/symeda/sormas/api/i18n/Captions.java
@@ -802,6 +802,10 @@ public interface Captions {
 	String FeatureConfiguration_enabled = "FeatureConfiguration.enabled";
 	String FeatureConfiguration_endDate = "FeatureConfiguration.endDate";
 	String firstName = "firstName";
+	String FollowUp_followUpUntil = "FollowUp.followUpUntil";
+	String FollowUp_person = "FollowUp.person";
+	String FollowUp_reportDate = "FollowUp.reportDate";
+	String FollowUp_uuid = "FollowUp.uuid";
 	String formatNumberOfVisitsFormat = "formatNumberOfVisitsFormat";
 	String formatNumberOfVisitsLongFormat = "formatNumberOfVisitsLongFormat";
 	String formatSimpleNumberFormat = "formatSimpleNumberFormat";
