+++ sormas-api/src/main/resources/captions.properties
@@ -908,6 +908,12 @@ formatNumberOfVisitsFormat=%d (%d missed)
 formatNumberOfVisitsLongFormat=%d visits (%d missed)
 formatSimpleNumberFormat=%d
 
+# FollowUp
+FollowUp.uuid=Follow-up ID
+FollowUp.person=Follow-up person
+FollowUp.reportDate=Date of report
+FollowUp.followUpUntil=Follow-up until
+
 # HealthConditions
 HealthConditions=Health conditions
 HealthConditions.tuberculosis=Tuberculosis
