+++ sormas-api/src/main/resources/captions_fr-FR.properties
@@ -906,6 +906,12 @@ formatNumberOfVisitsFormat=%d (%d manqué)
 formatNumberOfVisitsLongFormat=%d visites (%d manquées)
 formatSimpleNumberFormat=%d
 
+# FollowUp
+FollowUp.uuid=ID de suivi
+FollowUp.person=Personne de suivi
+FollowUp.reportDate=Date du signalement
+FollowUp.followUpUntil=Suivi jusqu'à
+
 # HealthConditions
 HealthConditions=Etat de santé
 HealthConditions.tuberculosis=Tuberculose
